backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
python-jose==3.3.0
authlib==1.3.0
redis>=5.0.1 